[FIX] FND-97 Rearrage folders, Breaking down profile page into components, fix linkedIn nullable
