+++ frontend/src/app/dashboard/discover/candidates/page.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 import React from "react";
 import image from "../../../assets/people.png";
-import ExperienceBoard from "@/components/profile/ExperienceBoard";
-import EducationBoard from "@/components/profile/EducationBoard";
+import ExperienceBoard from "@/components/discover/ExperienceBoard";
+import EducationBoard from "@/components/discover/EducationBoard";
 
 const Profile = () => {
   return (
