+++ frontend/src/components/discover/EducationBoard.tsx
