+++ frontend/src/components/discover/EducationTag.tsx
