+++ frontend/src/components/discover/ExperienceBoard.tsx
